pyproject.toml
[project]
name = "samantha-os1"
version = "0.1.0"
description = "Samantha OS1 is a conversational AI assistant."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "chainlit==2.0.dev0",
    "langchain-groq>=0.2.0",
    "langchain-openai>=0.2.2",
    "langchain>=0.3.3",
    "openai==1.51.2",
    "plotly>=5.24.1",
    "tavily-python>=0.5.0",
    "together>=1.3.1",
    "websockets==13.1",
    "yfinance>=0.2.44",
]
